% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.r
\name{add_rownames}
\alias{add_rownames}
\title{Set rownames of data.frame or matrix}
\usage{
add_rownames(df, rownames)
}
\arguments{
\item{df}{data.frame or matrix}

\item{rownames}{rownames to add}
}
\value{
df with rownames added
}
\description{
Set rownames of data.frame or matrix and return it, for use with pipes
}
