% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.r
\name{allDups}
\alias{allDups}
\title{All duplicates.}
\usage{
allDups(vec)
}
\arguments{
\item{vec}{A vector.}
}
\value{
A boolean vector of the same length as \code{vec}.  TRUE if the element is duplicated, FALSE if not.
}
\description{
Finds all duplicates in a vector including first instances of duplicates.
}
\examples{
allDups(sample(1:100, size=100, replace=T)
}
