% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.r
\name{append<-}
\alias{append<-}
\title{Append to a vector.}
\usage{
append(x) <- value
}
\arguments{
\item{x}{A vector.}

\item{value}{The element to append.}
}
\description{
Appends the 2nd argument to the 1st.
}
\examples{
x <- 1:5
append(x) <- 6
print(x)

[1] 1 2 3 4 5 6
}
