% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.r
\name{bioc}
\alias{bioc}
\title{Install Bioconductor package.}
\usage{
bioc(package)
}
\arguments{
\item{package}{unquoted package name.}
}
\description{
Utility function for installing packages from bioconductor easily.
}
\examples{
\code{bioc(DESeq2)}
}
