% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.r
\name{chop}
\alias{chop}
\title{Cleans leading and trailing whitespace.}
\usage{
chop(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A vector with leading and trailing whitespace removed.
}
\description{
Removes leading and trailing whitespace in a vector of strings.
}
\examples{
chop(c("    hello ", "  123  \t"))

[1] "hello" "123"
}
