% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions (1).r, R/my_functions.r
\name{cindex}
\alias{cindex}
\alias{cindex}
\title{Concordance Index.}
\usage{
cindex(probs, time, event)

cindex(probs, time, event)
}
\arguments{
\item{probs}{The prognostic score of each patient.}

\item{time}{The time of each patient.}

\item{event}{The death of each patient: 1 for a patient death, 0 for censored.}

\item{probs}{The prognostic score of each patient.}

\item{time}{The time of each patient.}

\item{event}{The death of each patient: 1 for a patient death, 0 for censored.}
}
\value{
The concordance index.

The concordance index.
}
\description{
Calculates the concordance index from the results of a censored survival model.  Very fast compared to other packages.

Calculates the concordance index from the results of a censored survival model.  Very fast compared to other packages.
}
