% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions (1).r, R/my_functions.r
\name{f1score}
\alias{f1score}
\alias{f1score}
\title{F1 score.}
\usage{
f1score(probs, class)

f1score(probs, class)
}
\arguments{
\item{probs}{A numeric vector where 1 is predicted positive and 0 is predicted negative.}

\item{class}{A numeric vector where 1 is positive and 0 is negative.}

\item{probs}{A numeric vector where 1 is predicted positive and 0 is predicted negative.}

\item{class}{A numeric vector where 1 is positive and 0 is negative.}
}
\value{
The F1 score.

The F1 score.
}
\description{
Calculates F1 score from the results of a classification model.

Calculates F1 score from the results of a classification model.
}
\seealso{
\url{https://en.wikipedia.org/wiki/F1_score}

\url{https://en.wikipedia.org/wiki/F1_score}
}
