% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.r
\name{fastAUC}
\alias{fastAUC}
\title{Fast AUC}
\usage{
fastAUC(probs, class)
}
\arguments{
\item{probs}{A numeric vector of probabilities or likelihoods for each data point.}

\item{class}{A numeric vector where 1 is positive and 0 is negative.}
}
\value{
The AUC.
}
\description{
This function calculates the Area Under the Reciever-Operator Curve from the results of a classifcation model.
}
\examples{
library(AppliedPredictiveModeling)
data(abalone)
library(glmnet)
class <- factor(as.numeric(abalone$Type == "M"))
data <- data.matrix(abalone[,-1])
fit <- cv.glmnet(data, class, family="binomial")
probs <- predict(fit, newx=data)[,1]
fastAUC(probs, class)
}
\seealso{
Reference https://stat.ethz.ch/pipermail/r-help/2005-September/079872.html
}
