% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions (1).r, R/my_functions.r
\name{fastPR}
\alias{fastPR}
\alias{fastPR}
\title{Precision-Recall curve}
\usage{
fastPR(probs, class)

fastPR(probs, class)
}
\arguments{
\item{probs}{A numeric vector of probabilities or likelihoods for each data point.}

\item{class}{A numeric vector where 1 is positive and 0 is negative.}

\item{probs}{A numeric vector of probabilities or likelihoods for each data point.}

\item{class}{A numeric vector where 1 is positive and 0 is negative.}
}
\value{
a list containing the ROC curve.

a list containing the ROC curve.
}
\description{
Calculates the points in a Precision-Recall from the results of a classifcation model.

Calculates the points in a Precision-Recall from the results of a classifcation model.
}
\examples{
library(AppliedPredictiveModeling)
data(abalone)
library(glmnet)
class <- factor(as.numeric(abalone$Type == "M"))
data <- data.matrix(abalone[,-1])
fit <- cv.glmnet(data, class, family="binomial")
probs <- predict(fit, newx=data)[,1]
with(fastPR(probs, class), plot(recall, precision, type="l"))
library(AppliedPredictiveModeling)
data(abalone)
library(glmnet)
class <- factor(as.numeric(abalone$Type == "M"))
data <- data.matrix(abalone[,-1])
fit <- cv.glmnet(data, class, family="binomial")
probs <- predict(fit, newx=data)[,1]
with(fastPR(probs, class), plot(recall, precision, type="l"))
}
