% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.r
\name{fastReadLines}
\alias{fastReadLines}
\title{Fast readLines.}
\usage{
fastReadLines(fname, newlinechar = "\\n", nchars = NULL)
}
\arguments{
\item{fname}{Filename to read.}

\item{newlinechar}{The new line character in the file.}
}
\value{
A vector containing all the lines in the file.
}
\description{
Replacement for readLines, faster.
}
\examples{
library(microbenchmark)
writeLines(replicate(100000, sample(letters, size=100, replace=T)), con="/tmp/temp.txt")
microbenchmark(fastReadLines("/tmp/temp.txt"), 
               readLines("/tmp/temp.txt"), times=3)[,c("expr", "mean"), drop=F]

Unit: milliseconds
                          expr      min       lq     mean   median       uq
fastReadLines("/tmp/temp.txt") 335.3874 335.9188 336.5900 336.4502 337.1912
    readLines("/tmp/temp.txt") 724.9136 725.4523 727.2444 725.9911 728.4098
}
