% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.r
\name{fpkmFromCounts}
\alias{fpkmFromCounts}
\title{Calculate FPKM.}
\usage{
fpkmFromCounts(mat, gene_lengths, uq_norm = T)
}
\arguments{
\item{mat}{An integer matrix representing the counts of a RNA-Seq dataset.}

\item{gene_lengths}{A named vector with the length of each gene.}

\item{uq_norm}{If TRUE, will perform upper-quartile normalization.}
}
\value{
A matrix containing FPKM with the same dimensions as mat.
}
\description{
Calculates FPKM from a count matrix.
}
\seealso{
FPKM-UQ	A normalized read count in which gene expression values, in FPKM, are divided by the 75th percentile value.
\url{https://gdc-docs.nci.nih.gov/Data/Bioinformatics_Pipelines/Expression_mRNA_Pipeline/}
\url{http://vinaykmittal.blogspot.com/2013/10/fpkmrpkm-normalization-caveat-and-upper.html}
}
