% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions (1).r, R/my_functions.r
\name{head2}
\alias{head2}
\alias{head2}
\title{Upper-left corner of matrix.}
\usage{
head2(x, n = 10, ncols = 10)

head2(x, n = 10, ncols = 10)
}
\arguments{
\item{x}{A wide matrix or data.frame.}

\item{n}{Number of lines to display.  Default 10.}

\item{ncols}{Number of columns to display.  Default 10.}

\item{x}{A wide matrix or data.frame.}

\item{n}{Number of lines to display.  Default 10.}

\item{ncols}{Number of columns to display.  Default 10.}
}
\value{
\code{n} by \code{ncols} subset of the matrix taken from the upper-left corner.

\code{n} by \code{ncols} subset of the matrix taken from the upper-left corner.
}
\description{
Shortcut function for previewing a matrix or data.frame by displaying the upper-left corner.  Similar to \code{head}.

Shortcut function for previewing a matrix or data.frame by displaying the upper-left corner.  Similar to \code{head}.
}
