% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions (1).r, R/my_functions.r
\name{install}
\alias{install}
\alias{install}
\title{Install CRAN package.}
\usage{
install(package, repos = "http://cran.us.r-project.org")

install(package, repos = "http://cran.us.r-project.org")
}
\arguments{
\item{package}{unquoted package name.}

\item{package}{unquoted package name.}
}
\description{
Utility function for installing packages from CRAN easily.

Utility function for installing packages from CRAN easily.
}
\examples{
\\code{install(Rcpp)}
\\code{install(Rcpp)}
}
