% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.r
\name{logit}
\alias{logit}
\title{Logit Transformation.}
\usage{
logit(x)
}
\arguments{
\item{x}{Input to the function (e.g., probabilities from 0 to 1).}
}
\value{
Logit(x)
}
\description{
Calculates the results of the Logit Transformation.
}
