% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions (1).r, R/my_functions.r
\name{matrixFactor}
\alias{matrixFactor}
\alias{matrixFactor}
\title{Matrix Factor Design.}
\usage{
matrixFactor(x, names = NULL)

matrixFactor(x, names = NULL)
}
\arguments{
\item{x}{A factor.}

\item{names}{The name of each instance in the resulting matrix (i.e., the rownames).}

\item{x}{A factor.}

\item{names}{The name of each instance in the resulting matrix (i.e., the rownames).}
}
\value{
The design matrix.

The design matrix.
}
\description{
From a factor, returns a design matrix with a column for each level.

From a factor, returns a design matrix with a column for each level.
}
\examples{
matrixFactor(factor(letters))
matrixFactor(factor(letters))
}
