% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions (1).r, R/my_functions.r
\name{mccscore}
\alias{mccscore}
\alias{mccscore}
\title{Matthew's correlation coefficient}
\usage{
mccscore(probs, class)

mccscore(probs, class)
}
\arguments{
\item{probs}{A numeric vector where 1 is predicted positive and 0 is predicted negative.}

\item{class}{A numeric vector where 1 is positive and 0 is negative.}

\item{probs}{A numeric vector where 1 is predicted positive and 0 is predicted negative.}

\item{class}{A numeric vector where 1 is positive and 0 is negative.}
}
\value{
The MCC score.

The MCC score.
}
\description{
Calculates Matthew's correlation coefficient.  Requires the Rmpfr package.

Calculates Matthew's correlation coefficient.  Requires the Rmpfr package.
}
\seealso{
\url{https://en.wikipedia.org/wiki/Matthews_correlation_coefficient}

\url{https://en.wikipedia.org/wiki/Matthews_correlation_coefficient}
}
