% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.r
\name{mcreadRDS}
\alias{mcreadRDS}
\title{Multi-threaded readRDS}
\usage{
mcreadRDS(file, mc.cores = min(parallel::detectCores(), 4))
}
\arguments{
\item{file}{The filename of the rds object.}

\item{mc.cores}{How many cores to use in pigz.  The program does not seem to benefit after more than about 4 cores.}
}
\value{
The R object.
}
\description{
Uses the pigz utility to improve loading large R objects.  This is compatible with saveRDS and readRDS functions.  Requires pigz (sudo apt-get install pigz on Ubuntu).
}
\examples{
x <- sample(1e4, 1e7, replace=T)
saveRDS(x, file="temp.Rds")
xmc <- mcreadRDS("temp.Rds")
}
\seealso{
\url{http://stackoverflow.com/questions/28927750/}
}
