% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.r
\name{mcsaveRDS}
\alias{mcsaveRDS}
\title{Multi-threaded saveRDS}
\usage{
mcsaveRDS(object, file, mc.cores = min(parallel::detectCores(), 4))
}
\arguments{
\item{object}{An r object to save.}

\item{file}{The filename to save to.}

\item{mc.cores}{How many cores to use in pigz.  The program does not seem to benefit after more than about 4 cores.}
}
\description{
Uses the pigz utility to improve saving large R objects.  This is compatible with saveRDS and readRDS functions.  Requires pigz (sudo apt-get install pigz on Ubuntu).
}
\examples{
x <- sample(1e4, 1e7, replace=T)
y <- sample(1e4, 1e7, replace=T)
microbenchmark(mcsaveRDS(x, file="temp.Rds"), saveRDS(y, file="temp2.Rds")

Unit: seconds
                            expr      min       lq     mean   median       uq
 mcsaveRDS(x, file = "temp.Rds") 1.908310 1.908310 1.908310 1.908310 1.908310
  saveRDS(y, file = "temp2.Rds") 6.271499 6.271499 6.271499 6.271499 6.271499
}
\seealso{
\url{http://stackoverflow.com/questions/28927750/what-is-the-method-to-save-objects-in-a-compressed-form-using-multiple-threads-c}
}
