% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.r
\name{mcsplitapply}
\alias{mcsplitapply}
\title{Parallel split-matrix loop.}
\usage{
mcsplitapply(mat, f, func, mc.cores = 4, .combine = rbind, ...)
}
\arguments{
\item{mat}{The matrix.}

\item{f}{A factor of length equal to nrow(mat).  The levels of this factor will split the matrix into subsets.}

\item{func}{The function to apply to each subset.}

\item{mc.cores}{The number of cores to use.}

\item{.combine}{The function to combine the results with.  Default is rbind.  Use NA to return a list.}
}
\value{
A list or a combined object depending on the .combine parameter.
}
\description{
Splits a matrix into subsets based on a factor, and applies a function to each subset.  Typical use case: sum exon count data to gene count data.
}
\examples{
library(pasilla)
library(DEXSeq)
library(trqwe)
data(pasillaDEXSeqDataSet)
exon_counts <- counts(dxd)
f <- rowData(dxd)$groupID
gene_counts <- mcsplitapply(exon_counts, f, colSums)
}
