% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.r
\name{mgrepl}
\alias{mgrepl}
\title{Multiple grepl.}
\usage{
mgrepl(patterns, x, ...)
}
\arguments{
\item{x}{A vector of strings to search.}

\item{pattern}{A vector of regex patterns.}
}
\value{
A vector of the same length as \code{x}, TRUE if any pattern matches.
}
\description{
Takes in a list of regex patterns and returns true if any pattern matches.
}
\examples{
x <- fastReadLines("http://textfiles.com/ufo/ufobooks.ufo", newlinechar="\\r\\n", 1e5)
head(x[mgrepl(c("ALIENS", "UFO"), x)])

[1] "                         UFOLOGY BOOKS (REVISION 2.1 343 books)"
[2] "     H. S. Stewart on the  subject of UFO's. The list is alphabetic"
[3] "     Tom Mickus's most excellent board UFONET I.  (416-237-1204)"
[4] "     Bill Adler               *   LETTERS TO THE AIR FORCE ON UFOS  1967"
[5] "     Gordon W. Allen              OVERLORDS OLYMPIANS AND THE UFO   1974"
[6] "     Robert B. Beard              FLYING SAUCERS, UFO'S AND EXTRA"
}
