% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions (1).r, R/my_functions.r
\name{nelson_aelen_surv}
\alias{nelson_aelen_surv}
\alias{nelson_aelen_surv}
\title{Nelson Aelen estimator}
\usage{
nelson_aelen_surv(time, event)

nelson_aelen_surv(time, event)
}
\arguments{
\item{time}{The time of each patient.}

\item{event}{The death of each patient: 1 for a patient death, 0 for censored.}

\item{time}{The time of each patient.}

\item{event}{The death of each patient: 1 for a patient death, 0 for censored.}
}
\value{
A list containing the cumulative hazard function.

A list containing the cumulative hazard function.
}
\description{
Nelson–Aalen estimator is an estimator of the cumulative hazard function in survival data.  It can be used to compare the overall risks of two groups or used to estimate the number of deaths before a certain time.

Nelson–Aalen estimator is an estimator of the cumulative hazard function in survival data.  It can be used to compare the overall risks of two groups or used to estimate the number of deaths before a certain time.
}
\examples{
library(survival)
data(veteran)
with(nelson_aelen_surv(veteran$time, veteran$status), plot(ti, Hi, type="b"))
library(survival)
data(veteran)
with(nelson_aelen_surv(veteran$time, veteran$status), plot(ti, Hi, type="b"))
}
\seealso{
#https://en.wikipedia.org/wiki/Nelson-Aalen_estimator

#https://en.wikipedia.org/wiki/Nelson-Aalen_estimator
}
