% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions (1).r, R/my_functions.r
\name{prepend<-}
\alias{prepend<-}
\alias{prepend<-}
\title{Prepend to a vector.}
\usage{
prepend(x) <- value

prepend(x) <- value
}
\arguments{
\item{x}{A vector.}

\item{value}{The element to append.}

\item{x}{A vector.}

\item{value}{The element to append.}
}
\description{
Prepends the 2nd argument to the 1st.

Prepends the 2nd argument to the 1st.
}
\examples{
x <- 1:5
prepend(x) <- 6
print(x)

[1] 6 1 2 3 4 5
x <- 1:5
prepend(x) <- 6
print(x)

[1] 6 1 2 3 4 5
}
