% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions (1).r, R/my_functions.r
\name{reload}
\alias{reload}
\alias{reload}
\title{Reload a package.}
\usage{
reload(package, pos = 2)

reload(package, pos = 2)
}
\arguments{
\item{package}{Unquoted package name.}

\item{pos}{Namespace search position.}

\item{package}{Unquoted package name.}

\item{pos}{Namespace search position.}
}
\description{
Unload and reload a package and sets the namespace search order.

Unload and reload a package and sets the namespace search order.
}
\examples{
\\code{reload(trqwe)}
\\code{reload(trqwe)}
}
