% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions (1).r, R/my_functions.r
\name{reloadtrqwe}
\alias{reloadtrqwe}
\alias{reloadtrqwe}
\title{Unload and reload trqwe.}
\usage{
reloadtrqwe()

reloadtrqwe()
}
\description{
Unload and reload trqwe.
Shortcut for \code{reload(trqwe)}

Unload and reload trqwe.
Shortcut for \code{reload(trqwe)}
}
