% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.r
\name{set_colnames}
\alias{set_colnames}
\title{Set colnames of data.frame or matrix}
\usage{
set_colnames(df, colnames)
}
\arguments{
\item{df}{data.frame or matrix}

\item{colnames}{colnames to add}
}
\value{
df with colnames added
}
\description{
Set colnames of data.frame or matrix and return it, for use with pipes
}
