% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions (1).r, R/my_functions.r
\name{set_rownames}
\alias{set_rownames}
\alias{set_rownames}
\title{Set rownames of data.frame or matrix}
\usage{
set_rownames(df, rownames)

set_rownames(df, rownames)
}
\arguments{
\item{df}{data.frame or matrix}

\item{rownames}{rownames to add}

\item{df}{data.frame or matrix}

\item{rownames}{rownames to add}
}
\value{
df with rownames added

df with rownames added
}
\description{
Set rownames of data.frame or matrix and return it, for use with pipes

Set rownames of data.frame or matrix and return it, for use with pipes
}
