% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.r
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid Function.}
\usage{
sigmoid(x)
}
\arguments{
\item{probs}{x Input to the function.}
}
\value{
Sigmoid(x)
}
\description{
Calculates the results of the sigmoid function.
}
