% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions (1).r, R/my_functions.r
\name{tablec}
\alias{tablec}
\alias{tablec}
\title{Fast C++ tabulation.}
\usage{
tablec(x, sort = F)

tablec(x, sort = F)
}
\arguments{
\item{x}{A character, integer or factor vector.  NAs are allowed.}

\item{sort}{TRUE if the result names should be sorted alphanumerically.}

\item{x}{A character, integer or factor vector.  NAs are allowed.}

\item{sort}{TRUE if the result names should be sorted alphanumerically.}
}
\value{
A integer vector of counts of each element.

A integer vector of counts of each element.
}
\description{
Takes in a character, integer or factor vector and tabulates the number of times each element appears.

Takes in a character, integer or factor vector and tabulates the number of times each element appears.
}
\examples{

x <- factor(sample(1e5, 1e8, replace=T))
microbenchmark(table(x), tablec(x), times=3)

Unit: milliseconds
     expr       min         lq       mean    median         uq        max
 table(x) 9777.6457 10479.0949 10717.3382 11180.544 11187.1844 11193.8246
tablec(x)  678.0364   685.9467   713.1181   693.857   730.6589   767.4608


x <- sample(letters, 1e8, replace=T)
microbenchmark(table(x), tablec(x), tablec(x,sort=T), times=3)

Unit: seconds
               expr      min       lq     mean   median       uq      max
           table(x) 5.778514 5.829125 5.855560 5.879737 5.894083 5.908430
          tablec(x) 1.589360 1.589381 1.589516 1.589402 1.589594 1.589786
tablec(x, sort = T) 1.589386 1.590520 1.591824 1.591655 1.593044 1.594432

x <- factor(sample(1e5, 1e8, replace=T))
microbenchmark(table(x), tablec(x), times=3)

Unit: milliseconds
     expr       min         lq       mean    median         uq        max
 table(x) 9777.6457 10479.0949 10717.3382 11180.544 11187.1844 11193.8246
tablec(x)  678.0364   685.9467   713.1181   693.857   730.6589   767.4608


x <- sample(letters, 1e8, replace=T)
microbenchmark(table(x), tablec(x), tablec(x,sort=T), times=3)

Unit: seconds
               expr      min       lq     mean   median       uq      max
           table(x) 5.778514 5.829125 5.855560 5.879737 5.894083 5.908430
          tablec(x) 1.589360 1.589381 1.589516 1.589402 1.589594 1.589786
tablec(x, sort = T) 1.589386 1.590520 1.591824 1.591655 1.593044 1.594432
}
