% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.r
\name{tail2}
\alias{tail2}
\title{Lower-left corner of matrix.}
\usage{
tail2(x, n = 10, ncols = 10)
}
\arguments{
\item{x}{A wide matrix or data.frame.}

\item{n}{Number of lines to display.  Default 10.}

\item{ncols}{Number of columns to display.  Default 10.}
}
\value{
\code{n} by \code{ncols} subset of the matrix taken from the lower-left corner.
}
\description{
Shortcut function for previewing the bottom of a matrix or data.frame by displaying the lower-left corner.  Similar to \code{tail}.
}
