% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.r
\name{TCGA_barcode}
\alias{TCGA_barcode}
\title{Parse TCGA barcode.}
\usage{
TCGA_barcode(x, what = "patient")
}
\arguments{
\item{x}{TCGA barcode, or a vector of barcodes.}

\item{what}{Which information to return.}
}
\value{
The specified information contained in the barcode of the same length as \code{x}.
}
\description{
Taking in a full TCGA sample barcode, or any subset of the barcode, and return extracted values.
}
\examples{
TCGA_barcode(c("TCGA-02-0001-01C-01D-0182-01", "TCGA-02-0001-11C-01D-0182-01"), what="tissue")

[1] "01" "11"
}
