% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.r
\name{tost_power}
\alias{tost_power}
\title{Calculates power/sample size of a TOST test}
\usage{
tost_power(mu_A, mu_B, sd_A, sd_B, delta = 0.3, kappa = 1, alpha = 0.05,
  power = 0.8, method = "equivalence", n_iterations = 1000, paired = F,
  sample_var = 0)
}
\arguments{
\item{mu_A}{mean of group A}

\item{mu_B}{mean of group B}

\item{sd_A}{SD of group A}

\item{sd_B}{SD of group B}

\item{delta}{maximum tolerated difference (i.e., abs(mu_A - mu_B))}

\item{kappa}{sample size of group A / sample size of group B (nA / nB)}

\item{alpha}{significance threshold (0.05 default)}

\item{power}{Desired power level (0.8 default)}

\item{method}{Whether to use the exact statistics or simulation.  "Exact" - exact statistics, "equivalence" - simulate using equivalence package tost function, "TOSTER" - simulate using TOSTER functions}

\item{n_iterations}{Number of simulation iterations}

\item{paired}{If the TOST is paired (kappa should be 1)}

\item{paired_var}{Sample variance for paired TOST}
}
\value{
For exact statistics, returns a list of two values: minimum sample size to achieve desired power and the power at that sample size.  For simulation, returns a series of sample sizes and achieved power up to the desired power.
}
\description{
Calculates the sample size required for a TOST test, given a sample size through exact statistics or simulation
}
\examples{
res_equivalence <- tost_power(mu_A=1, mu_B=1, sd_A=0.2, sd_B=0.2, delta=0.2, method="equivalence")
res_toster <- tost_power(mu_A=1, mu_B=1, sd_A=0.2, sd_B=0.2, delta=0.2, method="TOSTER")
res_exact <- tost_power(mu_A=1, mu_B=1, sd_A=0.2, sd_B=0.2, delta=0.2, method="exact")

plot(res_equivalence$nB, res_equivalence$Power, xlim=c(1,20), ylim=c(0,1), type="s", main="TOST power analysis")
lines(res_toster$nB, res_toster$Power, col="green", type="s")
abline(h=0.8, col="blue", lty=2)
points(res_exact$nB, res_exact$Power, col="red", pch=18, cex=2)
legend("topleft", legend=c("equivalence", "TOSTER", "exact"), col=c("black", "green", "red"), pch=1)

res_equivalence <- tost_power(mu_A=1, mu_B=1, sd_A=0.2, sd_B=0.2, delta=0.1, method="equivalence")
res_exact <- tost_power(mu_A=1, mu_B=1, sd_A=0.2, sd_B=0.2, delta=0.1, method="exact")
plot(res_equivalence$nB, res_equivalence$Power, xlim=c(1,80), ylim=c(0,1), type="s", main="TOST power analysis")
abline(h=0.8, col="blue", lty=2)
points(res_exact$nB, res_exact$Power, col="red", pch=18, cex=2)
legend("topleft", legend=c("equivalence", "exact"), col=c("black", "red"), pch=1)
}
\seealso{
\url{http://powerandsamplesize.com/Calculators/Compare-2-Means/2-Sample-Equality}
Chow S, Shao J, Wang H. 2008. Sample Size Calculations in Clinical Research. 2nd Ed. Chapman & Hall/CRC Biostatistics Series. page 58.
}
