% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions (1).r, R/my_functions.r
\name{TOSTtwo}
\alias{TOSTtwo}
\alias{TOSTtwo}
\title{#' Multiple return assignment
#' @description Python style multiple return assignment.  
#' @examples
#' mreturn[x,y,z] <- list("hello", c(1,2,3), sqrt(2))
#' print(x)
#' [1] "hello"
#' print(y)
#' [1] 1 2 3
#' print(z)
#' [1] 1.414214
#' @rdname mreturn
mreturn <- structure(NA, class = "mreturn")}
\usage{
TOSTtwo(m1, m2, sd1, sd2, n1, n2, low_eqbound_d, high_eqbound_d, alpha,
  var.equal)

TOSTtwo(m1, m2, sd1, sd2, n1, n2, low_eqbound_d, high_eqbound_d, alpha,
  var.equal)
}
\description{
`[<-.mreturn` <- function(mreturn, ..., value) {
  vars <- sapply(substitute(list(...)), deparse)[-1]
  stopifnot(all(make.names(vars) == vars)) # invalid variable names
  stopifnot(length(value) == length(vars)) # incorrect number of return values
  for(i in 1:length(vars)) {
    assign(vars[i], value=value[[i]], envir = parent.frame())
  }
  return(mreturn)
}

`[<-.mreturn` <- function(mreturn, ..., value) {
  vars <- sapply(substitute(list(...)), deparse)[-1]
  stopifnot(all(make.names(vars) == vars)) # invalid variable names
  stopifnot(length(value) == length(vars)) # incorrect number of return values
  for(i in 1:length(vars)) {
    assign(vars[i], value=value[[i]], envir = parent.frame())
  }
  return(mreturn)
}
}
