% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.r
\name{varSizes}
\alias{varSizes}
\title{Size of R objects.}
\usage{
varSizes(env = globalenv(), units = "KB")
}
\arguments{
\item{env}{The environment to search (default global environment).}

\item{units}{Units to print out for each variable.}
}
\value{
A data.frame containing the size of each object.
}
\description{
Prints out the size of all R objects in the environment.
}
