% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.r
\name{ww_test}
\alias{ww_test}
\title{Wald-Wolfowitz Runs Tests for Randomness}
\usage{
ww_test(x)
}
\arguments{
\item{x}{A vector of items, coerced into a factor.}
}
\value{
A p-value for the statistical test.
}
\description{
This is the k-category asymptotic Z Test with continuity correction.  Imagine rolling a die multiple times to obtain a sequence of rolls.  This statistic tests whether there exists a "run" within the sequence where one particular number comes up more times in a row than expected randomly.  If the test is significant, it can be concluded that the die rolls are not independent.
}
\examples{
set.seed(1)
ww_test(sample(2, 100, replace=T))
ww_test(c(sample(6, 90, replace=T),rep(1,10)))
}
\seealso{
Reference https://ncss-wpengine.netdna-ssl.com/wp-content/themes/ncss/pdf/Procedures/NCSS/Analysis_of_Runs.pdf
}
