% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions (1).r, R/my_functions.r
\name{cosineDist}
\alias{cosineDist}
\alias{cosineDist}
\title{Cosine distance.}
\usage{
cosineDist(x)

cosineDist(x)
}
\arguments{
\item{x}{A matrix.}

\item{x}{A matrix.}
}
\value{
Cosine distance as a dist object.

Cosine distance as a dist object.
}
\description{
Calculates the cosine distance of rows of a matrix.

Calculates the cosine distance of rows of a matrix.
}
\examples{
rbind(c(1,1,1,0,0,0), c(0,0,0,1,1,1), c(1,1,1,0,0,1)) -> x
cosineDist(x)
          1         2
2 1.0000000
3 0.1339746 0.7113249
rbind(c(1,1,1,0,0,0), c(0,0,0,1,1,1), c(1,1,1,0,0,1)) -> x
cosineDist(x)
          1         2
2 1.0000000
3 0.1339746 0.7113249
}
\seealso{
\url{http://stackoverflow.com/questions/2535234/find-cosine-similarity-in-r}

\url{http://stackoverflow.com/questions/2535234/find-cosine-similarity-in-r}
}
