% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.r
\name{\%Q\%}
\alias{\%Q\%}
\title{Concatenate strings.}
\usage{
a \%Q\% b
}
\arguments{
\item{a}{First string.}

\item{b}{Second string.}
}
\value{
The concatenated string.
}
\description{
Concatenates two strings.
}
\examples{
'Hello ' \%Q\% 'World'

[1] "Hello World"
}
