% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.r
\name{hal2Apply}
\alias{hal2Apply}
\title{Apply on server}
\usage{
hal2Apply(X, FUN, varlist = ls(envir), .packages = "auto",
  envir = parent.frame(), mc.preschedule = F, mc.cores.server = 8,
  mc.cores.local = 3, user = Sys.info()["user"], server_ip = "hal2",
  revtunnel = T, ...)
}
\description{
Using PSOCK cluster.  Like mclapply, but on a server.
}
\examples{
Datain <- data.frame(x=1:100)
z <- 1
FUN <- function(i) {
  Datain \%>\% summarize(n=n()) \%>\% {(.$n+z+i)^2}
}
ret1 <- hal2Apply(1:1e4, FUN, varlist=c("Datain", "z"), .packages="dplyr", mc.cores=8, mc.preschedule = T)
}
