% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.r
\name{make_percent}
\alias{make_percent}
\title{Make percentage from number}
\usage{
make_percent(x, digits = 2)
}
\arguments{
\item{x}{A number.}

\item{digits}{Number of decimal places, default 2.}
}
\value{
A percentage.
}
\description{
Make percentage from number
}
\examples{
make_percent(0.424, 2)
}
