% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions (1).r, R/my_functions.r
\name{se}
\alias{se}
\alias{se}
\title{Standard error.}
\usage{
se(x)

se(x)
}
\arguments{
\item{x}{A vector.}

\item{x}{A vector.}
}
\value{
The standard error of x.

The standard error of x.
}
\description{
Calculates the standard error of a sampling distribution.

Calculates the standard error of a sampling distribution.
}
\examples{
x <- rnorm(1e3)
se(x)
[1] 0.03192027
x <- rnorm(1e3)
se(x)
[1] 0.03192027
}
